{"schema":"decodex/commit/1","summary":"Split Decodex app login panel modules","authority":"manual"}
