+++ apps/decodex-app/Sources/DecodexApp/LoginButtons.swift
@@ -1,58 +1,5 @@
 import SwiftUI
 
-struct LoginCodeBoxesView: View {
-	let code: String
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		HStack(spacing: 4) {
-			ForEach(0..<boxCount, id: \.self) { index in
-				if index == 4 {
-					Spacer()
-						.frame(width: 2)
-				}
-
-				Text(character(at: index))
-					.font(LoginFont.code)
-					.monospacedDigit()
-					.foregroundStyle(code.isEmpty ? LoginPalette.secondaryText(colorScheme).opacity(0.42) : LoginPalette.primaryText(colorScheme))
-					.frame(width: 23, height: 31)
-					.background {
-						RoundedRectangle(cornerRadius: 7, style: .continuous)
-							.fill(LoginPalette.codeBoxFill(colorScheme))
-					}
-					.overlay {
-						RoundedRectangle(cornerRadius: 7, style: .continuous)
-							.strokeBorder(LoginPalette.codeBoxStroke(colorScheme), lineWidth: 0.75)
-							.allowsHitTesting(false)
-					}
-					.shadow(
-						color: LoginPalette.codeBoxShadow(colorScheme),
-						radius: colorScheme == .dark ? 2.5 : 1.4,
-						y: 0.7
-					)
-			}
-		}
-		.frame(maxWidth: .infinity, alignment: .center)
-	}
-
-	private var characters: [String] {
-		code.map { String($0).uppercased() }
-	}
-
-	private var boxCount: Int {
-		max(8, min(12, characters.count))
-	}
-
-	private func character(at index: Int) -> String {
-		guard characters.indices.contains(index) else {
-			return ""
-		}
-
-		return characters[index]
-	}
-}
-
 struct LoginIconActionButton: View {
 	let symbol: String
 	let feedbackSymbol: String?
@@ -162,7 +109,7 @@ struct LoginTextButtonStyle: ButtonStyle {
 			.modernGlassSurface(cornerRadius: 12, depth: .control)
 			.scaleEffect(configuration.isPressed ? 0.965 : 1)
 			.opacity(isEnabled ? 1 : 0.64)
-			.animation(.interactiveSpring(response: 0.16, dampingFraction: 0.78), value: configuration.isPressed)
+			.animation(PanelMotion.press, value: configuration.isPressed)
 	}
 
 	private var foreground: Color {
@@ -171,52 +118,5 @@ struct LoginTextButtonStyle: ButtonStyle {
 		}
 
 		return LoginPalette.secondaryText(colorScheme)
-	}
-
-}
-
-enum LoginPalette {
-	static func primaryText(_ colorScheme: ColorScheme) -> Color {
-		colorScheme == .dark
-			? Color(red: 0.98, green: 0.985, blue: 1).opacity(0.99)
-			: Color(red: 0.09, green: 0.11, blue: 0.15).opacity(0.96)
-	}
-
-	static func secondaryText(_ colorScheme: ColorScheme) -> Color {
-		colorScheme == .dark
-			? Color(red: 0.86, green: 0.89, blue: 0.95).opacity(0.94)
-			: Color(red: 0.28, green: 0.33, blue: 0.4).opacity(0.86)
-	}
-
-	static func accent(_ colorScheme: ColorScheme) -> Color {
-		PanelPalette.actionBlue(colorScheme)
-	}
-
-	static func warning(_ colorScheme: ColorScheme) -> Color {
-		PanelPalette.warning(colorScheme)
-	}
-
-	static func feedback(_ colorScheme: ColorScheme) -> Color {
-		colorScheme == .dark
-			? Color(red: 0.86, green: 0.93, blue: 1)
-			: Color(red: 0.13, green: 0.32, blue: 0.52)
-	}
-
-	static func codeBoxFill(_ colorScheme: ColorScheme) -> Color {
-		colorScheme == .dark
-			? Color(red: 0.08, green: 0.1, blue: 0.14).opacity(0.72)
-			: Color(red: 0.96, green: 0.975, blue: 1).opacity(0.92)
-	}
-
-	static func codeBoxStroke(_ colorScheme: ColorScheme) -> Color {
-		colorScheme == .dark
-			? Color.white.opacity(0.16)
-			: Color(red: 0.48, green: 0.55, blue: 0.64).opacity(0.3)
-	}
-
-	static func codeBoxShadow(_ colorScheme: ColorScheme) -> Color {
-		colorScheme == .dark
-			? Color.black.opacity(0.22)
-			: Color(red: 0.24, green: 0.32, blue: 0.42).opacity(0.08)
 	}
 }
